Remove about to start callback and simplify code
