+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/domain/CCDCase.java
@@ -143,6 +143,4 @@ public class CCDCase {
     private CCDContactPartyType contactChangeParty;
     private CCDContactChangeContent contactChangeContent;
     private LocalDate calculatedResponseDeadline;
-    private LocalDate calculatedServiceDate;
-    private LocalDate extendedResponseDeadline;
 }
