+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/domain/defendant/CCDRespondent.java
@@ -28,6 +28,7 @@ public class CCDRespondent {
     private String letterHolderId;
     private String defendantId;
     private LocalDate responseDeadline;
+    private LocalDate extendedResponseDeadline;
     private LocalDate servedDate;
 
     private CCDParty claimantProvidedDetail;
