+++ src/main/java/uk/gov/hmcts/cmc/claimstore/events/EventProducer.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.cmc.claimstore.events;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
 import uk.gov.hmcts.cmc.claimstore.events.ccj.CountyCourtJudgmentEvent;
 import uk.gov.hmcts.cmc.claimstore.events.ccj.InterlocutoryJudgmentEvent;
 import uk.gov.hmcts.cmc.claimstore.events.ccj.ReDeterminationEvent;
@@ -26,7 +27,6 @@ import uk.gov.hmcts.cmc.claimstore.events.solicitor.RepresentedClaimCreatedEvent
 import uk.gov.hmcts.cmc.claimstore.events.solicitor.RepresentedClaimIssuedEvent;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.offers.MadeBy;
-import uk.gov.hmcts.reform.sendletter.api.Document;
 
 import java.time.LocalDate;
 
@@ -122,8 +122,8 @@ public class EventProducer {
 
     public void createPaperDefenceEvent(
             Claim claim,
-            Document coverLetter,
-            Document oconForm
+            CCDDocument coverLetter,
+            CCDDocument oconForm
     ) {
         publisher.publishEvent(new PaperDefenceReadyToPrintEvent(claim, coverLetter, oconForm));
     }
