+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/IssuePaperDefenceForms/IssuePaperDefenceCallbackHandler.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker.IssuePaperDefenceForms;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
@@ -8,6 +10,7 @@ import uk.gov.hmcts.cmc.ccd.domain.CCDCollectionElement;
 import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
 import uk.gov.hmcts.cmc.ccd.domain.defendant.CCDRespondent;
+import uk.gov.hmcts.cmc.claimstore.events.EventProducer;
 import uk.gov.hmcts.cmc.claimstore.services.IssueDateCalculator;
 import uk.gov.hmcts.cmc.claimstore.services.ResponseDeadlineCalculator;
 import uk.gov.hmcts.cmc.claimstore.services.UserService;
@@ -16,6 +19,7 @@ import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
+import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker.ChangeContactDetailsPostProcessor;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
@@ -29,47 +33,47 @@ import java.util.Map;
 
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.ISSUE_PAPER_DEFENSE_FORMS;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CASEWORKER;
-import static uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.generalletter.GeneralLetterService.DRAFT_LETTER_DOC;
 
 @Service
 @ConditionalOnProperty("feature_toggles.ctsc_enabled")
 public class IssuePaperDefenceCallbackHandler extends CallbackHandler {
-    public static final String CALCULATED_RESPONSE_DEADLINE = "calculatedResponseDeadline";
-    public static final String CALCULATED_SERVICE_DATE = "calculatedServiceDate";
-    public static final String COVER_LETTER_DOC = "coverLetterDoc";
-    public static final String LETTER_NAME = "%s-issue-paper-form.pdf";
     private static final List<Role> ROLES = Collections.singletonList(CASEWORKER);
     private static final List<CaseEvent> EVENTS = Collections.singletonList(ISSUE_PAPER_DEFENSE_FORMS);
+    private static final Logger logger = LoggerFactory.getLogger(ChangeContactDetailsPostProcessor.class);
+    private static final String ERROR_MESSAGE =
+            "There was a technical problem. Nothing has been sent. You need to try again.";
 
     private final CaseDetailsConverter caseDetailsConverter;
     private final UserService userService;
     private final ResponseDeadlineCalculator responseDeadlineCalculator;
     private final IssueDateCalculator issueDateCalculator;
-    private final IssuePaperResponsePostProccessor issuePaperResponsePostProccessor;
     private final IssuePaperResponseLetterService issuePaperResponseLetterService;
+    private final EventProducer eventProducer;
+    private final IssuePaperResponseNotificationService issuePaperResponseNotificationService;
 
     @Autowired
     public IssuePaperDefenceCallbackHandler(
             CaseDetailsConverter caseDetailsConverter,
             UserService userService,
+            EventProducer eventProducer,
             ResponseDeadlineCalculator responseDeadlineCalculator,
             IssueDateCalculator issueDateCalculator,
-            IssuePaperResponsePostProccessor issuePaperResponsePostProccessor,
-            IssuePaperResponseLetterService issuePaperResponseLetterService
+            IssuePaperResponseLetterService issuePaperResponseLetterService,
+            IssuePaperResponseNotificationService issuePaperResponseNotificationService
     ) {
+        this.eventProducer = eventProducer;
         this.caseDetailsConverter = caseDetailsConverter;
         this.userService = userService;
         this.responseDeadlineCalculator = responseDeadlineCalculator;
         this.issueDateCalculator = issueDateCalculator;
-        this.issuePaperResponsePostProccessor = issuePaperResponsePostProccessor;
         this.issuePaperResponseLetterService = issuePaperResponseLetterService;
+        this.issuePaperResponseNotificationService = issuePaperResponseNotificationService;
     }
 
     @Override
     protected Map<CallbackType, Callback> callbacks() {
         return Map.of(
-                CallbackType.ABOUT_TO_START, this::updateDeadlinesAndGenerateDocuments,
-                CallbackType.ABOUT_TO_SUBMIT, issuePaperResponsePostProccessor::performPostProcesses
+                CallbackType.ABOUT_TO_SUBMIT, this::updateDeadlinesAndSendDocuments
         );
     }
 
@@ -84,29 +88,37 @@ public class IssuePaperDefenceCallbackHandler extends CallbackHandler {
     }
 
 
-    private AboutToStartOrSubmitCallbackResponse updateDeadlinesAndGenerateDocuments(CallbackParams callbackParams) {
+    private AboutToStartOrSubmitCallbackResponse updateDeadlinesAndSendDocuments(CallbackParams callbackParams) {
         CaseDetails caseDetails = callbackParams.getRequest().getCaseDetails();
         CCDCase ccdCase = caseDetailsConverter.extractCCDCase(caseDetails);
         Claim claim = caseDetailsConverter.extractClaim(caseDetails);
         String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
         var userDetails = userService.getUserDetails(authorisation);
         String caseworkerName = userDetails.getFullName();
+        CCDRespondent respondent = ccdCase.getRespondents().get(0).getValue();
 
         CCDCase updatedCCDCase = addNewDeadlinesToCCDCase(ccdCase);
 
-        CCDRespondent newRespondent = updatedCCDCase.getRespondents().get(0).getValue();
-        Map<String, Object> data = Map.of(CALCULATED_RESPONSE_DEADLINE, newRespondent.getResponseDeadline()
-                , CALCULATED_SERVICE_DATE, newRespondent.getServedDate());
-
-        CCDDocument coverLetter = issuePaperResponseLetterService.createCoverLetter(ccdCase, caseworkerName, authorisation);
-        CCDDocument oconForm = issuePaperResponseLetterService.createOconForm(ccdCase, claim, authorisation);
-
-        data.put(COVER_LETTER_DOC, coverLetter);
-        data.put(DRAFT_LETTER_DOC, oconForm);
-
-        return AboutToStartOrSubmitCallbackResponse.builder()
-                .data(data)
+        Claim updatedClaim = claim.toBuilder()
+                .responseDeadline(respondent.getResponseDeadline())
+                .serviceDate(respondent.getServedDate())
                 .build();
+
+        CCDDocument coverLetter = issuePaperResponseLetterService
+                .createCoverLetter(updatedCCDCase, caseworkerName, authorisation);
+        CCDDocument oconForm = issuePaperResponseLetterService.createOconForm(updatedCCDCase, claim, authorisation);
+
+
+        var builder = AboutToStartOrSubmitCallbackResponse.builder();
+        try {
+            issuePaperResponseNotificationService.notifyClaimant(updatedClaim);
+            eventProducer.createPaperDefenceEvent(updatedClaim, oconForm, coverLetter);
+            return builder.data(caseDetailsConverter.convertToMap(updatedCCDCase)).build();
+        } catch (Exception e) {
+            logger.error("Error notifying citizens", e);
+            return builder.errors(Collections.singletonList(ERROR_MESSAGE)).build();
+
+        }
     }
 
     private CCDCase addNewDeadlinesToCCDCase(CCDCase ccdCase) {
@@ -120,14 +132,13 @@ public class IssuePaperDefenceCallbackHandler extends CallbackHandler {
         CCDRespondent respondent = collectionElement.getValue().toBuilder()
                 .responseDeadline(newResponseDeadline)
                 .servedDate(newServiceDate)
+                .extendedResponseDeadline(newExtendedResponseDeadline)
                 .build();
-
         return ccdCase.toBuilder()
                 .respondents(List.of(CCDCollectionElement.<CCDRespondent>builder()
                         .value(respondent)
                         .id(collectionElement.getId())
                         .build()))
-                .extendedResponseDeadline(newExtendedResponseDeadline)
                 .build();
     }
